refactored api to PartsOffApi, added retrofit dependencies
