+++ app/src/main/java/com/dreamdevs/partsoff_app/SignUpActivity.kt
@@ -4,7 +4,7 @@ import android.os.Bundle
 import android.widget.Toast
 import androidx.appcompat.app.AppCompatActivity
 import androidx.appcompat.widget.Toolbar
-import com.dreamdevs.partsoff_app.api.RetrofitInstance
+import com.dreamdevs.partsoff_app.partsOffApi.RetrofitInstance
 import com.dreamdevs.partsoff_app.databinding.ActivitySignUpBinding
 import com.dreamdevs.partsoff_app.partsOffModels.authModels.RegisterResponse
 import retrofit2.Call
