--- app/src/main/java/com/dreamdevs/partsoff_app/UserAdapter.kt
@@ -1,49 +1,0 @@
-package com.dreamdevs.partsoff_app
-
-import android.view.LayoutInflater
-import android.view.ViewGroup
-import androidx.recyclerview.widget.AsyncListDiffer
-import androidx.recyclerview.widget.DiffUtil
-import androidx.recyclerview.widget.RecyclerView
-import com.dreamdevs.partsoff_app.databinding.ItemUserBinding
-import com.dreamdevs.partsoff_app.userclasses.*
-import java.util.Arrays
-
-class UserAdapter : RecyclerView.Adapter<UserAdapter.UserViewHolder>() {
-
-    inner class UserViewHolder(val binding: ItemUserBinding) : RecyclerView.ViewHolder(binding.root)
-
-    private val diffCallback = object : DiffUtil.ItemCallback<User>(){
-        override fun areItemsTheSame(oldItem: User, newItem: User): Boolean {
-            return oldItem.id == newItem.id
-        }
-
-        override fun areContentsTheSame(oldItem: User, newItem: User): Boolean {
-            return oldItem == newItem
-        }
-    }
-
-    private val differ = AsyncListDiffer(this, diffCallback)
-    var users: List<User>
-        get() = differ.currentList
-        set(value) {differ.submitList(value)}
-
-    override fun getItemCount() = users.size
-
-    override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): UserViewHolder {
-        return UserViewHolder(ItemUserBinding.inflate(
-            LayoutInflater.from(parent.context),
-            parent,
-            false
-        ))
-    }
-
-    override fun onBindViewHolder(holder: UserViewHolder, position: Int) {
-        holder.binding.apply {
-            val user = users[position]
-            val userData = arrayOf(user.id , user.name, user.username, user.email, user.phone, user.address)
-            tvTitle.text = userData.contentToString()
-        }
-    }
-}
-
